+++ database/migrations/2021_06_21_115122_create_在途量_table.php
@@ -17,7 +17,7 @@ class Create在途量Table extends Migration
             $table->string('客戶', 30);
             $table->string('料號', 12);
             $table->primary(['客戶','料號']);
-            $table->float('請購數量');
+            $table->double('請購數量');
             $table->timestamps();
             $table->softDeletes();
         });
